use stdint types in cluster module

+++ include/ecewo/cluster.h
@@ -2,6 +2,7 @@
 #define ECEWO_CLUSTER_H
 
 #include <stdbool.h>
+#include <stdint.h>
 #include "ecewo-config.h"
 
 #ifndef ECEWO_HAS_CLUSTER
@@ -10,19 +11,19 @@
 
 typedef struct
 {
-    int workers;
+    uint8_t workers;
     bool respawn;
-    void (*on_worker_start)(int worker_id);
-    void (*on_worker_exit)(int worker_id, int status);
+    void (*on_worker_start)(uint8_t  worker_id);
+    void (*on_worker_exit)(uint8_t  worker_id, int status);
 } Cluster;
 
-bool cluster_init(const Cluster *config, int base_port, int argc, char **argv);
-int cluster_get_port(void);
+bool cluster_init(const Cluster *config, uint16_t base_port, int argc, char **argv);
+uint16_t cluster_get_port(void);
 bool cluster_is_master(void);
 bool cluster_is_worker(void);
-int cluster_worker_id(void);
-int cluster_worker_count(void);
-int cluster_cpu_count(void);
+uint8_t cluster_worker_id(void);
+uint8_t cluster_worker_count(void);
+uint8_t cluster_cpu_count(void);
 void cluster_signal_workers(int signal);
 void cluster_wait_workers(void);
